Move search into own controller
